// src/app/page/master/event/event.interface.ts
import {EventEmitter} from '@angular/core';

export interface Event {
	kode?: any;
	id?: any;
	// kdDepartemen?: any;
	kdProfile?: string;
	namaEvent?: string;
	kdEventHead?: string;
	version?: string;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
	id_kode?: any;
	namaEventHead?: any;
}